fix: Resolve 422 validation error in pending processing endpoint
Fix JSON serialization issues in InvoiceProcessingList schema:
- Change extracted_total_value from Decimal to float for proper JSON serialization
- Ensure errors are properly cast to list[str]
- Convert Decimal values to float in endpoint before creating response
- Handle nullable total_value safely with explicit type conversion

This resolves the 422 Unprocessable Entity errors when frontend fetches pending processing records.

+++ apps/api/src/schemas/invoice_processing.py
@@ -97,7 +97,7 @@ class InvoiceProcessingList(BaseModel):
     image_count: int
     confidence_score: float = Field(default=0.0, ge=0.0, le=1.0)
     extracted_issuer_name: Optional[str] = None
-    extracted_total_value: Optional[Decimal] = None
+    extracted_total_value: Optional[float] = None
     extracted_issue_date: Optional[datetime] = None
     errors: list[str] = Field(default_factory=list)
     created_at: datetime
